LPU-0004 Rename package to avoid discrepancies
